pyproject.toml
[project]
name = "gitlab-compliance-checker"
version = "0.1.0"
description = "A Python tool to check GitLab project compliance with predefined rules."
authors = [
  { name = "Lokesh Manchala", email = "manchalalokesh2003@gmail.com" }
]
readme = "README.md"
license = { text = "GNU Affero General Public License v3.0" }
requires-python = ">=3.11"
dependencies = [
  "requests>=2.31.0,<3.0.0",
  "streamlit>=1.32.0,<2.0.0",
  "xlsxwriter>=3.2.0,<4.0.0",
  "python-gitlab==2.10.1",
  "python-dotenv>=1.0.1,<2.0.0",
  "pandas>=2.2.0,<3.0.0",
  "python-dateutil>=2.9.0,<3.0.0",
  "numpy>=1.26.0,<3.0.0",
  "pytz>=2024.1",
  "tzdata>=2024.1; sys_platform == 'win32'"
]

[project.optional-dependencies]
dev = [
  "ruff",
  "pytest",
  "requests-mock",
  "pre-commit"
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "C", "B"]
exclude = ["tests/*"]

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q"
testpaths = ["tests"]

[project.urls]
"Homepage" = "https://code.swecha.org/tools/gitlab-compliance-checker"
"Issue Tracker" = "https://code.swecha.org/tools/gitlab-compliance-checker/-/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
